pyproject.toml
[project]
name = "VinhoBot"
version = "0.1.0"
description = "In vino veritas"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.30.0",
    "chainlit>=2.7.1.1",
    "greenlet>=3.2.4",
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "langchain-core>=0.3.74",
    "langchain-groq>=0.3.7",
    "langchain-huggingface>=0.3.1",
    "langchain-ollama>=0.3.6",
    "langchain-postgres>=0.0.15",
    "psycopg[binary,pool]>=3.2.9",
    "sentence-transformers>=5.1.0",
]

[dependency-groups]
dev = [
    "ipykernel>=6.30.1",
    "ipywidgets>=8.1.7",
    "python-dotenv>=1.1.1",
    "rich>=14.1.0",
    "ruff>=0.12.8",
]
